backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.2
django-filter>=23.0
dj-database-url>=1.2.0
psycopg2-binary>=2.9
gunicorn>=20.1
